+++ src/pages/task/newTask/AddNewTask.tsx
@@ -1806,6 +1806,12 @@ const AddNewTask = () => {
                   <Input
                     disabled={couponUsedBtn[0]}
                     onChange={handleChangeCoupon}
+                    onKeyPress={(e) => {
+                      const allowedCharacters = /^[a-zA-Z0-9]+$/;
+                      if (!allowedCharacters.test(e.key)) {
+                        e.preventDefault();
+                      }
+                    }}
                     style={{
                       paddingRight: 0,
                       paddingTop: 0,